+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * [#180](https://github.com/antoinezanardi/werewolves-assistant-web/issues/180) - Role types.
 * [#182](https://github.com/antoinezanardi/werewolves-assistant-web/issues/182) - Accordions in About page for roles.
+* [#185](https://github.com/antoinezanardi/werewolves-assistant-web/issues/185) - Reset base rules button in game roles options.
 
 ### ♻️ Refactoring
 
@@ -31,13 +32,13 @@
 * `@vue/cli-plugin-vuex` updated to version `4.5.12`.
 * `@vue/cli-service` updated to version `4.5.12`.
 * `chai` updated to version `4.3.4`.
-* `eslint` updated to version `7.23.0`.
+* `eslint` updated to version `7.24.0`.
 * `eslint-plugin-vue` updated to version `7.8.0`.
-* `particles.vue` updated to version `2.8.0`.
+* `particles.vue` updated to version `2.9.1`.
 * `qs` updated to version `6.10.1`.
 * `sweetalert2` updated to version `10.16.0`.
 * `v-tooltip` updated to version `2.1.3`.
-* `vue-i18n` updated to version `8.24.2`.
+* `vue-i18n` updated to version `8.24.3`.
 * `vue-tour` updated to version `2.0.0`.
 
 ---